requirements.txt
chromadb>=0.6.0
openai>=1.0.0
open-clip-torch>=2.26.0
torch>=2.0.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
numpy>=1.24.0
pillow>=9.0.0
